Return the number of source bytes read from new decompress function
This makes the re-worked decompress function resilient to being
passed data beyond a frame end. On each invocation the function
will return the decompressed data and the number of bytes read.
The caller can then decide what to do if the number of bytes read
is less than the size of the compressed data passed to the function.