Initialize destination storage on creation
If the LZ4 library writes 0 bytes, then the memory could remain uninitialized.
This caused a problem with a chunked roundtrip on the data b'0'. During de-
compression no bytes are written to the destination. Calling a Py_SIZE(dest)=0
produces a bytes object b'', but when this is compared with b'', it returns
False (i.e. dest == b'' returns false, even though printing dest produces b'').

This is a quirk of the Python C API. Or possibly a bug. Should a zero length
bytes object always be equal to b'', even if its storage is not initialized to
'\0' ?