Deploy from AppVeyor if on a tag irrespective of which branch.
Unfortunately github sets base_ref to be the first branch
a commit was seen on. So, merging to master without a merge
commit and then tagging means that base_ref will be set to
the branch the last commit was originally on, and not master.
This is all a bit flakey.