Add Python 3.9 builds on Appveyor
This required a shift to using the Visual Studio 2019 platform, as that's the
only image that includes Python 3.9. This resulted in a significant refactor of
the AppVeyor build.

We now specify the compiler to be used for each Python version in order to
follow: https://wiki.python.org/moin/WindowsCompilers.

In addition, we no longer need to use the build.cmd script, and so remove that.